* lib/install-sh: Fix typo in previous patch for handling --. Use more-consistent style for ';;'. Prefer || to if-then-else-:. * tests/install2.test: Rework to avoid set -e problems.
2006-12-24  Akim Demaille  <akim@epita.fr>

Simplify install-sh and its test.  This shouldn't change any behavior.
(This patch is a subset of the patch proposed in
<http://lists.gnu.org/archive/html/automake-patches/2006-10/msg00077.html>.)
* lib/install-sh (usage): Use usual GNU style.
(dstarg): Rename as...
(dst_arg): this for consistency.
Simplify quoting of assignments.
Sort them.
Don't use '\' to continue commands: && suffices.
Remove useless "continue" in the argument processing,
and factor the shifts.
* tests/defs.in: Some improvements to make it set -e clean.
Use the traditional ":" trick to protect loops from being empty.
Remove an empty straightforward piece of code prepared to define
additional variables.
Use test instead of [], for consistency.
* tests/install2.test: Use set -e, to simplify code.